Move decodeEntities to useShippingAddress
